Formally support missing values in the data/values fields.
This uses an attribute to signify the placeholder value used for missing values
(as we can't just assume that R's missingness policy applies here).

This also requires some rewording for other uses of the word 'missing' (namely,
for absent list elements) to avoid confusion with the concept of missing values.