Formally support missing values in the value and data fields.  (#2)
This uses an attribute to signify the placeholder value used for missing values
(as we can't just assume that R's missingness policy applies here). This is the
same approach used in, e.g., uzuki2 and other ArtifactDB schemas.

We perform some rewording for other uses of the word 'missing' (namely,
for absent list elements) to avoid confusion with the concept of missing values.