Switch to stricter datatype constraints for the various datasets. (#7)
This adds stricter type constraints to every HDF5 dataset. Now we specify the
largest possible type that is allowed; we don't just allow a datatype of the
right class, because that doesn't provide much actionable guarantees.

This library now has a dependency on ritsuko, allowing us to take advantage of
the various version parsing and HDF5 utilities there.

Missing value placeholders for numeric datasets now must be of the exact same
type as the dataset itself. This avoids casting-related discrepancies and is
consistent with the behavior of other ritsuko-dependent libraries.

We moved the specification into Markdown files for easier reading and to reduce
the scope of doxygen, which just concerns the C++ validators.

We deprecated the external HDF5 array, which never made much sense given that
we have no guarantee that the external array is saved like a dense array or
sparse matrix; better to use a custom array type instead, where the application
is responsible for offering such guarantees.

We now prohibit string seeds in the unary logic and combine operations.

Finally, we expanded tests to check for all versions at once.